requirements.txt
jupyter==1.0.0
Keras==2.4.3
matplotlib==3.0.3
notebook==6.0.1
numpy==1.18.2
Pillow==6.2.0
tensorflow==2.3.1
tqdm==4.36.1
